feat: support multiple profiles with an environment variable
This introduces the `RBW_PROFILE` environment variable to be able to run
multiple instances in parallel, that will use different configuration
paths.
To use this feature, simply run `rbw` with the environment variable set
to any string that is valid for a path. Setting it to an empty string is
like not setting it at all.
e.g.:
* `RBW_PROFILE=perso rbw list`
* `RBW_PROFILE=pro rbw list`